SAC-LSTM-MindSpore/requirements.txt
absl-py == 1.2.0
asttokens == 2.0.8
astunparse == 1.6.3
attrs == 22.1.0
certifi == 2022.6.15
contextlib2 == 21.6.0
cycler == 0.11.0
easydict == 1.9
fonttools == 4.37.1
imageio == 2.21.1
iniconfig == 1.1.1
joblib == 1.1.0
JPype1 == 1.4.0
kiwisolver == 1.4.4
matplotlib == 3.5.3
mindspore-gpu == 1.8.1
mindvision == 0.1.0
ml-collections == 0.1.1
networkx == 2.8.6
numpy == 1.23.2
opencv-contrib-python-headless == 4.6.0.66
opencv-python == 4.6.0.66
opencv-python-headless == 4.6.0.66
packaging == 21.3
Pillow == 9.2.0
pip == 22.1.2
pluggy == 1.0.0
protobuf == 4.21.5
psutil == 5.9.1
py == 1.11.0
pyparsing == 3.0.9
pytest == 7.1.3
python-dateutil == 2.8.2
PyWavelets == 1.3.0
PyYAML == 6.0
scikit-image == 0.19.3
scikit-learn == 1.1.2
scipy == 1.9.1
setuptools == 63.4.1
six == 1.16.0
thop == 0.1.1.post2207130030
threadpoolctl == 3.1.0
tifffile == 2022.8.12
tomli == 2.0.1
torch == 1.12.1
tqdm == 4.64.1
typing_extensions == 4.3.0
wheel == 0.37.1